Add files via upload
The parchment image is what I'm using for the background. Because of how Android Studio works you have to lay multiples of the image to fit the screen, the width is wrap content and the height to fit it is 295dp. You will need 3 to fit the whole screen. The other is the image I am using for the welcome button on the Welcome screen